python/requirements.txt
fhir.resources
requests
rich
inquirer
